+++ libnavui-dropin/src/main/java/com/mapbox/navigation/dropin/ViewOptionsCustomization.kt
@@ -39,6 +39,15 @@ class ViewOptionsCustomization {
      */
     var routeArrowOptions: RouteArrowOptions? = null
 
+    /**
+     * Set if the Info Panel should be visible for the Free Drive state. When set to `true`
+     * [InfoPanelBinder] will only render an empty info panel. You are responsible to inflate
+     * appropriate [Binder] to render the content you would like to show on the info panel during
+     * free drive.
+     * Set to `false` for the default behavior.
+     */
+    var showInfoPanelInFreeDrive: Boolean? = null
+
     companion object {
         /**
          * Default route line options.