+++ libnavui-dropin/src/main/java/com/mapbox/navigation/dropin/coordinator/InfoPanelCoordinator.kt
@@ -3,12 +3,14 @@ package com.mapbox.navigation.dropin.coordinator
 import android.view.View
 import android.view.ViewGroup
 import androidx.constraintlayout.widget.Guideline
+import androidx.core.view.ViewCompat
 import com.google.android.material.bottomsheet.BottomSheetBehavior
 import com.mapbox.navigation.base.ExperimentalPreviewMapboxNavigationAPI
 import com.mapbox.navigation.core.MapboxNavigation
 import com.mapbox.navigation.dropin.NavigationViewContext
 import com.mapbox.navigation.dropin.binder.infopanel.InfoPanelBinder
 import com.mapbox.navigation.dropin.binder.infopanel.InfoPanelHeaderBinder
+import com.mapbox.navigation.ui.app.internal.navigation.NavigationState
 import com.mapbox.navigation.ui.base.lifecycle.UIBinder
 import com.mapbox.navigation.ui.base.lifecycle.UICoordinator
 import kotlinx.coroutines.flow.Flow
@@ -30,6 +32,7 @@ internal class InfoPanelCoordinator(
     private val behavior = BottomSheetBehavior.from(infoPanel)
 
     init {
+        infoPanel.addOnLayoutChangeListener(FixBottomSheetLayoutWhenHidden(infoPanel, behavior))
         behavior.hide()
     }
 
@@ -38,9 +41,8 @@ internal class InfoPanelCoordinator(
 
         behavior.addBottomSheetCallback(updateGuideline)
         coroutineScope.launch {
-            store.select { it.destination }.collect { destination ->
-                if (destination != null) behavior.collapse()
-                else behavior.hide()
+            bottomSheetVisibility().collect { isVisible ->
+                updateBottomSheetVisibility(isVisible)
 
                 // When BottomSheet is already in requested state, BottomSheetCallback won't be
                 // called leaving guideline in a wrong position.
@@ -66,6 +68,18 @@ internal class InfoPanelCoordinator(
         }
     }
 
+    private fun bottomSheetVisibility() = combine(
+        store.select { it.navigation },
+        context.options.showInfoPanelInFreeDrive
+    ) { navigationState, showInfoPanelInFreeDrive ->
+        showInfoPanelInFreeDrive || navigationState != NavigationState.FreeDrive
+    }
+
+    private fun updateBottomSheetVisibility(isVisible: Boolean) {
+        if (isVisible) behavior.collapse()
+        else behavior.hide()
+    }
+
     private fun <V : View> BottomSheetBehavior<V>.collapse() {
         state = BottomSheetBehavior.STATE_COLLAPSED
         isHideable = false
@@ -88,6 +102,32 @@ internal class InfoPanelCoordinator(
 
         override fun onSlide(bottomSheet: View, slideOffset: Float) {
             setGuidelinePosition(bottomSheet)
+        }
+    }
+
+    /**
+     * An OnLayoutChangeListener that ensures the bottom sheet is always laid out at the bottom of
+     * the parent view when in STATE_HIDDEN.
+     */
+    private class FixBottomSheetLayoutWhenHidden(
+        private val layout: ViewGroup,
+        private val behavior: BottomSheetBehavior<ViewGroup>
+    ) : View.OnLayoutChangeListener {
+
+        override fun onLayoutChange(
+            v: View?,
+            left: Int,
+            top: Int,
+            right: Int,
+            bottom: Int,
+            oldLeft: Int,
+            oldTop: Int,
+            oldRight: Int,
+            oldBottom: Int
+        ) {
+            if (behavior.state == BottomSheetBehavior.STATE_HIDDEN) {
+                ViewCompat.offsetTopAndBottom(layout, (layout.parent as? View)?.height ?: 0)
+            }
         }
     }
 }